requirements.txt
numpy
scipy
opencv-python
pyaudio
pillow
darkdetect
keras
tensorflow
dlib
playsound
customtkinter==5.2.2
keras2ascii